+++ lib/arbitrum-verifier/src/lib.rs
@@ -1,11 +1,11 @@
 use core::fmt::Debug;
 
+use arbitrum_light_client_types::{ClientState, Header};
 use ethereum_verifier::verify::{verify_account_storage_root, verify_storage_proof};
 use sha3::{Digest, Keccak256};
 use unionlabs::{
     ethereum::slot::{MappingKey, Slot},
     hash::H256,
-    ibc::lightclients::arbitrum::{client_state::ClientState, header::Header},
     uint::U256,
 };
