--- lib/unionlabs/src/ibc/lightclients/arbitrum.rs
@@ -1,4 +1,0 @@
-pub mod client_state;
-pub mod consensus_state;
-pub mod header;
-pub mod l2_header;
