+++ light-clients/arbitrum-light-client/src/errors.rs
@@ -1,3 +1,4 @@
+use arbitrum_light_client_types::{ClientState, ConsensusState};
 use ethereum_light_client::errors::{CanonicalizeStoredValueError, InvalidCommitmentKey};
 use ics008_wasm_client::IbcClientError;
 use unionlabs::{
@@ -37,18 +38,9 @@ pub enum Error {
     //     >,
     // ),
     #[error("unable to decode client state")]
-    ClientStateDecode(
-        #[source]
-        DecodeErrorOf<Proto, unionlabs::ibc::lightclients::arbitrum::client_state::ClientState>,
-    ),
+    ClientStateDecode(#[source] DecodeErrorOf<Proto, ClientState>),
     #[error("unable to decode consensus state")]
-    ConsensusStateDecode(
-        #[source]
-        DecodeErrorOf<
-            Proto,
-            unionlabs::ibc::lightclients::arbitrum::consensus_state::ConsensusState,
-        >,
-    ),
+    ConsensusStateDecode(#[source] DecodeErrorOf<Proto, ConsensusState>),
 
     // REVIEW: Move this variant to IbcClientError?
     #[error("consensus state not found at height {0}")]
