+++ light-clients/arbitrum-light-client/types/Cargo.toml
@@ -1,0 +1,19 @@
+[package]
+edition = "2021"
+name    = "arbitrum-light-client-types"
+version = "0.1.0"
+
+[dependencies]
+protos         = { workspace = true, features = ["union+ibc+lightclients+arbitrum+v1"] }
+rlp            = { workspace = true, features = ["derive"] }
+serde          = { workspace = true, features = ["derive"] }
+sha3.workspace = true
+thiserror      = { workspace = true }
+unionlabs      = { workspace = true }
+
+[features]
+default = ["proto"]
+proto   = ["unionlabs/proto"]
+
+[dev-dependencies]
+hex-literal = { workspace = true }
