+++ light-clients/arbitrum-light-client/types/src/lib.rs
@@ -1,0 +1,8 @@
+pub mod client_state;
+pub mod consensus_state;
+pub mod header;
+pub mod l2_header;
+
+pub use crate::{
+    client_state::ClientState, consensus_state::ConsensusState, header::Header, l2_header::L2Header,
+};
